requirements.txt
streamlit==1.28.0
newspaper3k==0.2.8
transformers==4.34.0
torch==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
lxml==4.9.3
python-dotenv==1.0.0
fake-useragent==1.2.1